+++ OpenRA.Mods.Common/Traits/BotModules/Squads/States/AirStates.cs
@@ -140,7 +140,7 @@ namespace OpenRA.Mods.Common.Traits.BotModules.Squads
 
 			var activity = a.CurrentActivity;
 			var type = activity.GetType();
-			if (type == typeof(Resupply) || type == typeof(ResupplyAircraft))
+			if (type == typeof(Resupply))
 				return true;
 
 			var next = activity.NextActivity;
@@ -148,7 +148,7 @@ namespace OpenRA.Mods.Common.Traits.BotModules.Squads
 				return false;
 
 			var nextType = next.GetType();
-			if (nextType == typeof(Resupply) || nextType == typeof(ResupplyAircraft))
+			if (nextType == typeof(Resupply))
 				return true;
 
 			return false;
